intento de hacer baja logica de luces automaticas

+++ src/main/java/com/unla/grupo6/servicies/ILucesAutoService.java
@@ -5,13 +5,15 @@ import java.util.List;
 import com.unla.grupo6.entities.DisLucesAuto;
 import com.unla.grupo6.models.DisLucesAutoModel;
 
+
 public interface ILucesAutoService {
 
+	
 	public List<DisLucesAuto> getAll();
 
 	public void save(DisLucesAuto disLucesAuto);
-
-	public boolean remove(int id);
+	
+	public boolean remove(long id);
 
 	public DisLucesAuto buscar(long id);
